+++ pmd-core/src/test/java/net/sourceforge/pmd/processor/PmdRunnableTest.java
@@ -38,6 +38,7 @@ import net.sourceforge.pmd.lang.LanguageVersion;
 import net.sourceforge.pmd.lang.ast.Node;
 import net.sourceforge.pmd.lang.ast.Parser;
 import net.sourceforge.pmd.lang.ast.RootNode;
+import net.sourceforge.pmd.lang.document.FileId;
 import net.sourceforge.pmd.lang.impl.SimpleLanguageModuleBase;
 import net.sourceforge.pmd.lang.rule.AbstractRule;
 import net.sourceforge.pmd.util.ContextedAssertionError;
@@ -74,7 +75,7 @@ class PmdRunnableTest {
         configuration.setForceLanguageVersion(lv);
         configuration.setIgnoreIncrementalAnalysis(true);
         try (PmdAnalysis pmd = PmdAnalysis.create(configuration)) {
-            pmd.files().addSourceFile("foo", "test.dummy");
+            pmd.files().addSourceFile(FileId.fromPathLikeString("test.dummy"), "foo");
             pmd.addRuleSet(RuleSet.forSingleRule(rule));
             return pmd.performAnalysisAndCollectReport();
         }