+++ pmd-core/src/test/java/net/sourceforge/pmd/reporting/GlobalAnalysisListenerTest.java
@@ -28,6 +28,7 @@ import net.sourceforge.pmd.cache.AnalysisCache;
 import net.sourceforge.pmd.cache.NoopAnalysisCache;
 import net.sourceforge.pmd.lang.ast.FileAnalysisException;
 import net.sourceforge.pmd.lang.ast.Node;
+import net.sourceforge.pmd.lang.document.FileId;
 
 class GlobalAnalysisListenerTest {
 
@@ -76,7 +77,7 @@ class GlobalAnalysisListenerTest {
         MyFooRule rule = new MyFooRule();
         runPmd(config, GlobalAnalysisListener.noop(), rule);
 
-        verify(mockCache).checkValidity(any(), any());
+        verify(mockCache).checkValidity(any(), any(), any());
         verify(mockCache, times(1)).persist();
         verify(mockCache, times(NUM_DATA_SOURCES)).isUpToDate(any());
     }
@@ -92,7 +93,7 @@ class GlobalAnalysisListenerTest {
         runPmd(config, GlobalAnalysisListener.noop(), rule);
 
         // cache methods are called regardless
-        verify(mockCache).checkValidity(any(), any());
+        verify(mockCache).checkValidity(any(), any(), any());
         verify(mockCache, times(1)).persist();
         verify(mockCache, times(NUM_DATA_SOURCES)).isUpToDate(any());
     }
@@ -111,10 +112,10 @@ class GlobalAnalysisListenerTest {
             runPmd(config, listener, rule);
         });
 
-        assertEquals("fname1.dummy", exception.getFileName());
+        assertEquals("fname1.dummy", exception.getFileId().getOriginalPath());
 
         // cache methods are called regardless
-        verify(mockCache).checkValidity(any(), any());
+        verify(mockCache).checkValidity(any(), any(), any());
         verify(mockCache, times(1)).persist();
         verify(mockCache, times(1)).isUpToDate(any());
     }
@@ -131,9 +132,9 @@ class GlobalAnalysisListenerTest {
     private void runPmd(PMDConfiguration config, GlobalAnalysisListener listener, Rule rule) {
         try (PmdAnalysis pmd = PmdAnalysis.create(config)) {
             pmd.addRuleSet(RuleSet.forSingleRule(rule));
-            pmd.files().addSourceFile("abc", "fname1.dummy");
-            pmd.files().addSourceFile("abcd", "fname2.dummy");
-            pmd.files().addSourceFile("abcd", "fname21.dummy");
+            pmd.files().addSourceFile(FileId.fromPathLikeString("fname1.dummy"), "abc");
+            pmd.files().addSourceFile(FileId.fromPathLikeString("fname2.dummy"), "abcd");
+            pmd.files().addSourceFile(FileId.fromPathLikeString("fname21.dummy"), "abcd");
             pmd.addListener(listener);
             pmd.performAnalysis();
         }
@@ -144,7 +145,7 @@ class GlobalAnalysisListenerTest {
 
         @Override
         public void apply(Node node, RuleContext ctx) {
-            if (node.getTextDocument().getDisplayName().contains("1")) {
+            if (node.getTextDocument().getFileId().getFileName().contains("1")) {
                 ctx.addViolation(node);
             }
         }